requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
einops==0.8.1
et_xmlfile==2.0.0
faiss-cpu==1.13.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.48
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
openai==2.8.1
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pre_commit==4.5.0
propcache==0.4.1
pyarrow==22.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ragas==0.3.9
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.7.0
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
ty==0.0.1a28
typer==0.9.4
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
